feat: ensure product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -2,15 +2,15 @@
 
 import { auth } from "@/lib/auth";
 
-import { headers } from "next/headers";
 import { db } from "@/db";
-import { cartItemTable, cartTable } from "@/db/schema";
+import { cartItemTable } from "@/db/schema";
 import { eq } from "drizzle-orm";
-import { removeProductFromCartSchema } from "./schema";
+import { headers } from "next/headers";
 import z from "zod";
+import { DecreaseCartProductQuantitySchema, decreaseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (data: z.infer<typeof removeProductFromCartSchema>) => {
-  removeProductFromCartSchema.safeParse(data);
+export const decreaseCartProductQuantity = async (data: DecreaseCartProductQuantitySchema) => {
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -31,12 +31,17 @@ export const removeProductFromCart = async (data: z.infer<typeof removeProductFr
   if (!cartItem) {
     throw new Error("Product variant not found in cart");
   }
-  
-  const cartDoesNotBelongToUser = cartItem?.cart.userId !== session.user.id;
+
+  const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id;
 
   if (cartDoesNotBelongToUser) {
     throw new Error("User not authorized to remove this item");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if(cartItem.quantity == 1){
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  await db.update(cartItemTable).set({ quantity: cartItem.quantity - 1 }).where(eq(cartItemTable.id, cartItem.id));
 };
